more replacing of size_t with MPI_Offset
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@523 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/lib/header.c
@@ -130,7 +130,7 @@ ncmpii_NC_computeshapes(NC *ncp)
         }
  
         assert(ncp->begin_var > 0);
-        assert(ncp->xsz <= (size_t)ncp->begin_var);
+        assert(ncp->xsz <= ncp->begin_var);
         assert(ncp->begin_rec > 0);
         assert(ncp->begin_var <= ncp->begin_rec);
